test: address lifecycle isolation review findings
