+++ src/lib/actions/sandbox/connect-flow.test.ts
@@ -306,158 +306,6 @@ describe("connectSandbox flow", () => {
     );
   });
 
-  it("skips the vLLM model preflight only for probe-only connects (#4585)", async () => {
-    const harness = createConnectHarness();
-
-    await expect(harness.connectSandbox("alpha", { probeOnly: true })).resolves.toBeUndefined();
-    await expect(harness.connectSandbox("alpha", { probeOnly: true })).resolves.toBeUndefined();
-    expect(harness.preflightVllmSpy).not.toHaveBeenCalled();
-
-    await expect(harness.connectSandbox("alpha")).rejects.toThrow("process.exit(0)");
-    expect(harness.preflightVllmSpy).toHaveBeenCalledOnce();
-  });
-
-  it("warns and aligns a diverged route during a quiet probe-only connect (#3726)", async () => {
-    const harness = createConnectHarness({
-      inferenceGetOutput:
-        "Gateway inference:\n  Provider: nvidia-prod\n  Model: nvidia/nemotron-3-super-120b-a12b\n",
-      registryEntry: {
-        model: "claude-sonnet-4-20250514",
-        provider: "anthropic-prod",
-      },
-    });
-
-    await expect(harness.connectSandbox("alpha", { probeOnly: true })).resolves.toBeUndefined();
-
-    const errorOutput = harness.errorSpy.mock.calls.map((call) => String(call[0] ?? "")).join("\n");
-    expect(errorOutput).toContain("differs from the recorded route");
-    expect(errorOutput).toContain(
-      "Aligning the gateway to anthropic-prod/claude-sonnet-4-20250514",
-    );
-    expect(harness.runOpenshellSpy).toHaveBeenCalledWith(
-      [
-        "inference",
-        "set",
-        "--provider",
-        "anthropic-prod",
-        "--model",
-        "claude-sonnet-4-20250514",
-        "--no-verify",
-      ],
-      expect.objectContaining({ ignoreError: true }),
-    );
-    expect(harness.spawnSyncSpy).not.toHaveBeenCalledWith(
-      "openshell",
-      ["sandbox", "connect", "alpha"],
-      expect.any(Object),
-    );
-  });
-
-  it("wires the forced VM DNS monkeypatch into connect route repair", async () => {
-    vi.stubEnv("NEMOCLAW_FORCE_VM_DNS_MONKEYPATCH", "1");
-    const harness = createConnectHarness({
-      inferenceGetOutput:
-        "Gateway inference:\n  Provider: nvidia-prod\n  Model: nvidia/nemotron-3-super-120b-a12b\n",
-      inferenceProbeResponses: ['BROKEN 503 {"error":"inference service unavailable"}', "OK 200"],
-      registryEntry: {
-        model: "nvidia/nemotron-3-super-120b-a12b",
-        openshellDriver: "vm",
-        provider: "nvidia-prod",
-      },
-    });
-
-    await expect(harness.connectSandbox("alpha", { probeOnly: true })).resolves.toBeUndefined();
-
-    expect(harness.applyVmDnsMonkeypatchSpy).toHaveBeenCalledWith(
-      "alpha",
-      expect.objectContaining({ openshellDriver: "vm" }),
-    );
-    expect(harness.runSetupDnsProxySpy).not.toHaveBeenCalled();
-    expect(harness.runOpenshellSpy).not.toHaveBeenCalled();
-    const routeProbeCalls = harness.captureOpenshellSpy.mock.calls.filter((call) =>
-      JSON.stringify(call[0]).includes("inference.local/v1/models"),
-    );
-    expect(routeProbeCalls).toHaveLength(2);
-  });
-
-  it.each([
-    ["null", null, null],
-    ["provider-only", "nvidia-prod", null],
-    ["model-only", null, "nvidia/test"],
-    ["blank-provider", "   ", "nvidia/test"],
-    ["blank-model", "nvidia-prod", "   "],
-  ] as const)("skips inference reconciliation for %s registry entries (#5937)", async (_description, provider, model) => {
-    const harness = createConnectHarness({ registryEntry: { model, provider } });
-
-    await expect(harness.connectSandbox("alpha", { probeOnly: true })).resolves.toBeUndefined();
-
-    expect(harness.captureOpenshellSpy).not.toHaveBeenCalledWith(
-      ["inference", "get"],
-      expect.any(Object),
-    );
-    expect(harness.runOpenshellSpy).not.toHaveBeenCalled();
-  });
-
-  it("does not reset an inference route that already matches the sandbox", async () => {
-    const harness = createConnectHarness({
-      inferenceGetOutput:
-        "Gateway inference:\n  Provider: nvidia-prod\n  Model: nvidia/nemotron-3-super-120b-a12b\n",
-      registryEntry: {
-        model: "nvidia/nemotron-3-super-120b-a12b",
-        provider: "nvidia-prod",
-      },
-    });
-
-    await expect(harness.connectSandbox("alpha", { probeOnly: true })).resolves.toBeUndefined();
-
-    expect(harness.captureOpenshellSpy).toHaveBeenCalledWith(
-      ["inference", "get"],
-      expect.objectContaining({ ignoreError: true }),
-    );
-    expect(harness.runOpenshellSpy).not.toHaveBeenCalled();
-  });
-
-  it("stops before opening SSH when route repair and reset both fail", async () => {
-    const harness = createConnectHarness({
-      inferenceGetOutput:
-        "Gateway inference:\n  Provider: nvidia-prod\n  Model: nvidia/nemotron-3-super-120b-a12b\n",
-      inferenceProbeResponses: Array(7).fill('BROKEN 503 {"error":"upstream unavailable"}'),
-      registryEntry: {
-        model: "nvidia/nemotron-3-super-120b-a12b",
-        openshellDriver: "kubernetes",
-        provider: "nvidia-prod",
-      },
-    });
-
-    await expect(harness.connectSandbox("alpha")).rejects.toThrow("process.exit(1)");
-
-    expect(harness.runSetupDnsProxySpy).toHaveBeenCalledOnce();
-    expect(harness.runOpenshellSpy).toHaveBeenCalledOnce();
-    expect(harness.runOpenshellSpy).toHaveBeenCalledWith(
-      [
-        "inference",
-        "set",
-        "--provider",
-        "nvidia-prod",
-        "--model",
-        "nvidia/nemotron-3-super-120b-a12b",
-        "--no-verify",
-      ],
-      expect.objectContaining({ ignoreError: true }),
-    );
-    expect(harness.spawnSyncSpy).not.toHaveBeenCalledWith(
-      "openshell",
-      ["sandbox", "connect", "alpha"],
-      expect.any(Object),
-    );
-    const errorOutput = harness.errorSpy.mock.calls.map((call) => String(call[0] ?? "")).join("\n");
-    expect(errorOutput).toContain("inference.local is still unavailable");
-    expect(errorOutput).toContain(
-      "Connect is stopping because the sandbox inference route is known to be broken",
-    );
-    expect(exitSpy).toHaveBeenCalledWith(1);
-  });
-
   it("stops before opening SSH when the sandbox list reports a terminal failure phase", async () => {
     const harness = createConnectHarness({ listOutput: "alpha Error" });
