+++ src/lib/actions/sandbox/status-flow.test.ts
@@ -1,217 +1,12 @@
 // SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-import { createRequire } from "node:module";
-
 import { afterEach, beforeEach, describe, expect, it, type MockInstance, vi } from "vitest";
 
-import type { ProviderHealthStatus } from "../../inference/health";
-import type { SandboxGatewayState } from "./gateway-state";
-import type { SandboxStatusPreflightResult } from "./status-preflight";
-
-type ShowSandboxStatus = typeof import("./status")["showSandboxStatus"];
-
-const requireDist = createRequire(import.meta.url);
-const statusModulePath = "./status.js";
-
-// Warm the CommonJS source graph outside the first test's timeout. Each harness
-// still reloads the entry module after installing its dependency spies.
-requireDist(statusModulePath);
-delete require.cache[requireDist.resolve(statusModulePath)];
-
-type StatusFlowHarness = {
-  checkAgentVersionSpy: MockInstance;
-  collectSandboxStatusSnapshotSpy: MockInstance;
-  getActiveSandboxSessionsSpy: MockInstance;
-  getSandboxDockerRuntimeSpy: MockInstance;
-  logSpy: MockInstance;
-  removeSandboxSpy: MockInstance;
-  showSandboxStatus: ShowSandboxStatus;
-};
-
-type StatusFlowHarnessOptions = {
-  currentModel?: string;
-  currentProvider?: string;
-  inferenceHealth?: ProviderHealthStatus | null;
-  lookup?: SandboxGatewayState;
-  lookupState?: "present" | "missing";
-  preflight?: SandboxStatusPreflightResult;
-  processRunning?: boolean | null;
-  sandboxEntry?: Partial<Omit<typeof baseSandboxEntry, "agentVersion">> & {
-    agent?: string | null;
-    agentVersion?: string | null;
-  };
-  shieldsPosture?: {
-    mode: "locked" | "mutable_default" | "mutable";
-    detail: string;
-  };
-  versionCheck?: {
-    sandboxVersion?: string | null;
-    expectedVersion?: string | null;
-    isStale: boolean;
-    detectionMethod?: string;
-    schemeMismatch?: boolean;
-    verificationFailed?: boolean;
-  };
-};
-
-const baseSandboxEntry = {
-  name: "alpha",
-  model: "nvidia/nemotron",
-  provider: "ollama-local",
-  policies: ["npm", "telegram"],
-  hostGpuDetected: true,
-  gpuEnabled: true,
-  sandboxGpuEnabled: true,
-  sandboxGpuMode: "auto",
-  sandboxGpuDevice: "all",
-  sandboxGpuProof: {
-    status: "failed",
-    label: "cuInit",
-    detail: "CUDA initialization failed",
-  },
-  openshellDriver: "docker",
-  openshellVersion: "0.1.2",
-  dashboardPort: 18789,
-  agentVersion: "0.1.0",
-};
-
-function createStatusFlowHarness(options: StatusFlowHarnessOptions = {}) {
-  delete require.cache[requireDist.resolve(statusModulePath)];
-
-  const logSpy = vi.spyOn(console, "log").mockImplementation(() => undefined);
-  vi.spyOn(console, "error").mockImplementation(() => undefined);
-
-  const statusPreflight = requireDist("./status-preflight.js");
-  const statusSnapshot = requireDist("./status-snapshot.js");
-  const dockerHealth = requireDist("./docker-health.js");
-  const processRecovery = requireDist("./process-recovery.js");
-  const resolve = requireDist("../../adapters/openshell/resolve.js");
-  const agentRuntime = requireDist("../../agent/runtime.js");
-  const nim = requireDist("../../inference/nim.js");
-  const sandboxVersion = requireDist("../../sandbox/version.js");
-  const shields = requireDist("../../shields/index.js");
-  const registry = requireDist("../../state/registry.js");
-  const sandboxSession = requireDist("../../state/sandbox-session.js");
-
-  const lookup: SandboxGatewayState =
-    options.lookup ??
-    (options.lookupState === "missing"
-      ? {
-          state: "missing",
-          output: "sandbox alpha not found",
-          recoveredGateway: true,
-          recoveryVia: "gateway reattach",
-        }
-      : {
-          state: "present",
-          output: "Name: alpha\nPhase: Ready\nEndpoint: http://127.0.0.1:18789\n",
-          recoveredGateway: true,
-          recoveryVia: "gateway reattach",
-          recoveredSandbox: true,
-          recoverySandboxVia: "docker unpause",
-        });
-
-  const sandboxEntry = { ...baseSandboxEntry, ...options.sandboxEntry };
-
-  vi.spyOn(registry, "getSandbox").mockReturnValue(sandboxEntry);
-  const removeSandboxSpy = vi.spyOn(registry, "removeSandbox").mockImplementation(() => undefined);
-  vi.spyOn(statusPreflight, "getSandboxStatusPreflight").mockResolvedValue(
-    options.preflight ?? {
-      failure: null,
-      failureLayer: null,
-      suppressInferenceProbe: false,
-      exitCode: 0,
-    },
-  );
-  const collectSandboxStatusSnapshotSpy = vi
-    .spyOn(statusSnapshot, "collectSandboxStatusSnapshot")
-    .mockResolvedValue({
-      sb: sandboxEntry,
-      lookup,
-      rpcIssue: null,
-      currentModel: options.currentModel ?? "nvidia/nemotron-live",
-      currentProvider: options.currentProvider ?? "ollama-local",
-      inferenceHealth:
-        options.inferenceHealth === undefined
-          ? {
-              ok: true,
-              probed: true,
-              providerLabel: "Ollama",
-              endpoint: "http://127.0.0.1:11434/v1/chat/completions",
-              detail: "chat completions probe passed",
-              subprobes: [
-                {
-                  ok: false,
-                  probed: true,
-                  providerLabel: "Inference gateway chain",
-                  endpoint: "http://127.0.0.1:19000/v1/chat/completions",
-                  detail: "gateway refused connection",
-                  probeLabel: "gateway",
-                  failureLabel: "unreachable",
-                },
-              ],
-            }
-          : options.inferenceHealth,
-    });
-  const getSandboxDockerRuntimeSpy = vi
-    .spyOn(dockerHealth, "getSandboxDockerRuntime")
-    .mockReturnValue({
-      containerName: "openshell-alpha",
-      health: "unhealthy",
-      paused: false,
-    });
-  vi.spyOn(processRecovery, "isSandboxGatewayRunningForStatus").mockResolvedValue(
-    options.processRunning ?? false,
-  );
-  vi.spyOn(resolve, "resolveOpenshell").mockReturnValue("/usr/bin/openshell");
-  vi.spyOn(agentRuntime, "getSessionAgent").mockReturnValue({ name: "openclaw" });
-  vi.spyOn(agentRuntime, "getAgentDisplayName").mockReturnValue("OpenClaw");
-  vi.spyOn(agentRuntime, "getGatewayCommand").mockReturnValue("openclaw daemon");
-  vi.spyOn(nim, "nimStatus").mockReturnValue({
-    running: true,
-    healthy: false,
-    container: "alpha-nim",
-  });
-  vi.spyOn(nim, "nimStatusByName").mockReturnValue({
-    running: false,
-    healthy: false,
-    container: null,
-  });
-  vi.spyOn(nim, "shouldShowNimLine").mockReturnValue(true);
-  const checkAgentVersionSpy = vi.spyOn(sandboxVersion, "checkAgentVersion").mockReturnValue(
-    options.versionCheck ?? {
-      sandboxVersion: "0.1.0",
-      expectedVersion: "0.2.0",
-      isStale: true,
-      detectionMethod: "runtime",
-    },
-  );
-  vi.spyOn(shields, "getShieldsPosture").mockReturnValue(
-    options.shieldsPosture ?? {
-      mode: "mutable_default",
-      detail: "mutable default",
-    },
-  );
-  const getActiveSandboxSessionsSpy = vi
-    .spyOn(sandboxSession, "getActiveSandboxSessions")
-    .mockReturnValue({
-      detected: true,
-      sessions: [{ pid: 1 }, { pid: 2 }],
-    });
-
-  logSpy.mockClear();
-
-  return {
-    checkAgentVersionSpy,
-    collectSandboxStatusSnapshotSpy,
-    getActiveSandboxSessionsSpy,
-    getSandboxDockerRuntimeSpy,
-    logSpy,
-    removeSandboxSpy,
-    showSandboxStatus: requireDist(statusModulePath).showSandboxStatus,
-  } satisfies StatusFlowHarness;
-}
+import {
+  createStatusFlowHarness,
+  resetStatusFlowModuleCache,
+} from "../../../../test/support/status-flow-test-harness";
 
 describe("showSandboxStatus flow", () => {
   let exitSpy: MockInstance;
@@ -226,7 +21,7 @@ describe("showSandboxStatus flow", () => {
   afterEach(() => {
     vi.restoreAllMocks();
     process.exitCode = undefined;
-    delete require.cache[requireDist.resolve(statusModulePath)];
+    resetStatusFlowModuleCache();
   });
 
   it("prints the live sandbox, inference, runtime, session, version, and recovery signals", async () => {
