+++ src/lib/registry-recovery-seeded-paths.test.ts
@@ -1,0 +1,213 @@
+// SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+import { beforeEach, describe, expect, it, vi } from "vitest";
+
+import type { SandboxEntry } from "./state/registry.js";
+
+interface MockRegistryState {
+  sandboxes: Record<string, SandboxEntry>;
+  defaultSandbox: string | null;
+}
+
+const mockRegistryState: MockRegistryState = { sandboxes: {}, defaultSandbox: null };
+
+vi.mock("./state/registry.js", () => ({
+  listSandboxes: () => ({
+    sandboxes: Object.values(mockRegistryState.sandboxes),
+    defaultSandbox: mockRegistryState.defaultSandbox,
+  }),
+  getSandbox: (name: string) => mockRegistryState.sandboxes[name] ?? null,
+  registerSandbox: (entry: SandboxEntry) => {
+    mockRegistryState.sandboxes[entry.name] = entry;
+  },
+  updateSandbox: (name: string, partial: Partial<SandboxEntry>) => {
+    mockRegistryState.sandboxes[name] = {
+      ...mockRegistryState.sandboxes[name],
+      ...partial,
+    } as SandboxEntry;
+  },
+  setDefault: (name: string) => {
+    mockRegistryState.defaultSandbox = name;
+  },
+}));
+
+vi.mock("./adapters/openshell/resolve.js", () => ({
+  resolveOpenshell: vi.fn(),
+}));
+
+vi.mock("./gateway-runtime-action.js", () => ({
+  recoverNamedGatewayRuntime: vi.fn(),
+  getNamedGatewayLifecycleState: vi.fn(),
+}));
+
+vi.mock("./adapters/openshell/runtime.js", () => ({
+  captureOpenshell: vi.fn(),
+}));
+
+vi.mock("./state/onboard-session.js", () => ({
+  loadSession: vi.fn(),
+}));
+
+vi.mock("./runtime-recovery.js", () => ({
+  parseLiveSandboxEntries: vi.fn(),
+}));
+
+vi.mock("./runner.js", () => ({
+  validateName: (name: string) =>
+    /^[a-z]([a-z0-9-]*[a-z0-9])?$/.test(name)
+      ? name
+      : (() => {
+          throw new Error(`Invalid sandbox name: '${name}'`);
+        })(),
+}));
+
+import { resolveOpenshell } from "./adapters/openshell/resolve.js";
+import { captureOpenshell } from "./adapters/openshell/runtime.js";
+import {
+  getNamedGatewayLifecycleState,
+  recoverNamedGatewayRuntime,
+} from "./gateway-runtime-action.js";
+import { recoverRegistryEntries } from "./registry-recovery-action.js";
+import { parseLiveSandboxEntries } from "./runtime-recovery.js";
+import { loadSession } from "./state/onboard-session.js";
+
+const gammaEntry = (policies: string[]): SandboxEntry => ({
+  name: "gamma",
+  provider: "existing-provider",
+  model: "existing-model",
+  gpuEnabled: false,
+  policies,
+});
+
+const completedSession = (sandboxName: string, policyPresets: string[]) =>
+  ({
+    sandboxName,
+    provider: "nvidia-prod",
+    model: "nvidia/nemotron-3-super-120b-a12b",
+    policyPresets,
+    nimContainer: null,
+    steps: {
+      sandbox: { status: "complete", startedAt: null, completedAt: null, error: null },
+    },
+  }) as never;
+
+function resetSeededRecoveryMocks(): void {
+  mockRegistryState.sandboxes = {};
+  mockRegistryState.defaultSandbox = null;
+  vi.mocked(loadSession).mockReset().mockReturnValue(null);
+  vi.mocked(resolveOpenshell).mockReset().mockReturnValue("/usr/bin/openshell");
+  vi.mocked(recoverNamedGatewayRuntime)
+    .mockReset()
+    .mockResolvedValue({ recovered: true } as never);
+  vi.mocked(getNamedGatewayLifecycleState)
+    .mockReset()
+    .mockReturnValue({ state: "missing_named" } as never);
+  vi.mocked(captureOpenshell)
+    .mockReset()
+    .mockReturnValue({ output: "live sandboxes", status: 0 } as never);
+  vi.mocked(parseLiveSandboxEntries).mockReset().mockReturnValue([]);
+}
+
+describe("recoverRegistryEntries seeded recovery paths", () => {
+  beforeEach(resetSeededRecoveryMocks);
+
+  it("merges a confirmed session and additional live sandboxes without replacing the default", async () => {
+    mockRegistryState.sandboxes.gamma = gammaEntry(["npm"]);
+    mockRegistryState.defaultSandbox = "gamma";
+    vi.mocked(loadSession).mockReturnValue(completedSession("alpha", ["pypi"]));
+    vi.mocked(parseLiveSandboxEntries).mockReturnValue([
+      { name: "alpha", phase: "Ready" },
+      { name: "beta", phase: "Ready" },
+    ]);
+
+    const result = await recoverRegistryEntries();
+
+    expect(result.recoveredFromSession).toBe(true);
+    expect(result.recoveredFromGateway).toBe(1);
+    expect(result.sandboxes.map((sandbox) => sandbox.name).sort()).toEqual([
+      "alpha",
+      "beta",
+      "gamma",
+    ]);
+    expect(mockRegistryState.sandboxes.alpha?.policies).toEqual(["pypi"]);
+    expect(mockRegistryState.defaultSandbox).toBe("gamma");
+  });
+
+  it("skips invalid session and live sandbox names during seeded recovery", async () => {
+    mockRegistryState.sandboxes.gamma = gammaEntry([]);
+    mockRegistryState.defaultSandbox = "gamma";
+    vi.mocked(loadSession).mockReturnValue(completedSession("Alpha", []));
+    vi.mocked(parseLiveSandboxEntries).mockReturnValue([
+      { name: "alpha", phase: "Ready" },
+      { name: "Bad_Name", phase: "Ready" },
+    ]);
+
+    const result = await recoverRegistryEntries();
+
+    expect(result.sandboxes.map((sandbox) => sandbox.name).sort()).toEqual(["alpha", "gamma"]);
+    expect(mockRegistryState.sandboxes.Alpha).toBeUndefined();
+    expect(mockRegistryState.sandboxes.Bad_Name).toBeUndefined();
+    expect(mockRegistryState.defaultSandbox).toBe("gamma");
+  });
+
+  it("treats an incomplete (phantom) session as unseeded — stays in read-only/display-only path", async () => {
+    // PRA-2: a session that recorded sandboxName but whose sandbox step never
+    // completed is a phantom (#2753). It must NOT count as a recovery seed,
+    // otherwise an empty registry + phantom session would take the mutating,
+    // persisting seeded path. Recovery must stay read-only/display-only.
+    vi.mocked(loadSession).mockReturnValue({
+      sandboxName: "phantom",
+      provider: "nvidia",
+      model: "nemotron",
+      policyPresets: [],
+      nimContainer: null,
+      steps: {
+        sandbox: { status: "pending", startedAt: null, completedAt: null, error: null },
+      },
+    } as never);
+    vi.mocked(getNamedGatewayLifecycleState).mockReturnValue({ state: "healthy_named" } as never);
+    vi.mocked(parseLiveSandboxEntries).mockReturnValue([{ name: "dcode-station", phase: "Ready" }]);
+
+    const result = await recoverRegistryEntries();
+
+    // Read-only path: never invokes the mutating gateway recovery, inspects
+    // lifecycle directly, and surfaces the live sandbox display-only.
+    expect(recoverNamedGatewayRuntime).not.toHaveBeenCalled();
+    expect(getNamedGatewayLifecycleState).toHaveBeenCalledWith(undefined, {
+      ignoreProbeErrors: true,
+    });
+    const recovered = result.sandboxes.find((s) => s.name === "dcode-station") as
+      | { recoveredFromGateway?: boolean }
+      | undefined;
+    expect(recovered?.recoveredFromGateway).toBe(true);
+    // Nothing persisted — neither the phantom session sandbox nor the recovered one.
+    expect(mockRegistryState.sandboxes["dcode-station"]).toBeUndefined();
+    expect(mockRegistryState.sandboxes["phantom"]).toBeUndefined();
+  });
+
+  it("persists a requested live sandbox and makes it the default", async () => {
+    vi.mocked(parseLiveSandboxEntries).mockReturnValue([{ name: "alpha", phase: "Ready" }]);
+
+    const result = await recoverRegistryEntries({ requestedSandboxName: "alpha" });
+
+    expect(recoverNamedGatewayRuntime).toHaveBeenCalledOnce();
+    expect(getNamedGatewayLifecycleState).not.toHaveBeenCalled();
+    expect(result.recoveredFromGateway).toBe(1);
+    expect(result.sandboxes.map((sandbox) => sandbox.name)).toEqual(["alpha"]);
+    expect(mockRegistryState.sandboxes.alpha).toBeDefined();
+    expect(mockRegistryState.defaultSandbox).toBe("alpha");
+  });
+
+  it("keeps a missing requested sandbox absent while recovering other live entries", async () => {
+    vi.mocked(parseLiveSandboxEntries).mockReturnValue([{ name: "alpha", phase: "Ready" }]);
+
+    const result = await recoverRegistryEntries({ requestedSandboxName: "beta" });
+
+    expect(result.recoveredFromGateway).toBe(1);
+    expect(result.sandboxes.map((sandbox) => sandbox.name)).toEqual(["alpha"]);
+    expect(mockRegistryState.sandboxes.alpha).toBeDefined();
+    expect(mockRegistryState.sandboxes.beta).toBeUndefined();
+    expect(mockRegistryState.defaultSandbox).toBeNull();
+  });
+});
